docs(SteamClient): move shared enums/interfaces to a separate file
