+++ src/deck-components/steam-client/Downloads.ts
@@ -1,4 +1,4 @@
-import {Unregisterable} from "./index";
+import { Unregisterable } from "./shared";
 import {EAppUpdateError} from "./App";
 
 /**
