+++ src/deck-components/steam-client/Music.ts
@@ -1,4 +1,4 @@
-import {Unregisterable} from "./index";
+import { Unregisterable } from "./shared";
 
 /**
  * Represents functions related to controlling music playback in the Steam client.
