+++ src/deck-components/steam-client/OpenVR.ts
@@ -1,4 +1,4 @@
-import {Unregisterable} from "./index";
+import { Unregisterable } from "./shared";
 
 
 export interface OpenVR {
