+++ src/deck-components/steam-client/ServerBrowser.ts
@@ -1,4 +1,4 @@
-import {OperationResponse, Unregisterable} from "./index";
+import { OperationResponse, Unregisterable } from "./shared";
 
 /**
  * Represents functionality for the server browser.
