+++ src/deck-components/steam-client/URL.ts
@@ -1,4 +1,4 @@
-import { Unregisterable } from "./index";
+import { Unregisterable } from "./shared";
 
 export interface URL {
     /**
