+++ src/deck-components/steam-client/system/AudioDevice.ts
@@ -1,4 +1,4 @@
-import {JsPbMessage, Unregisterable} from "../index";
+import { JsPbMessage, Unregisterable } from "../shared";
 
 export interface AudioDevice {
     /**
