+++ src/deck-components/steam-client/system/Display.ts
@@ -1,4 +1,4 @@
-import {Unregisterable} from "../index";
+import { Unregisterable } from "../shared";
 
 export interface Display {
     EnableUnderscan(param0: boolean): any;
