+++ src/deck-components/steam-client/system/DisplayManager.ts
@@ -1,4 +1,4 @@
-import {Unregisterable} from "../index";
+import { Unregisterable } from "../shared";
 
 // CMsgSystemDisplayManagerState, CMsgSystemDisplayManagerSetMode
 export interface DisplayManager {
