+++ src/deck-components/steam-client/system/Perf.ts
@@ -1,4 +1,5 @@
-import {JsPbMessage, Unregisterable} from "../index";
+import { JsPbMessage } from "../shared";
+import { Unregisterable } from "../shared";
 
 // CMsgSystemPerfUpdateSettings, CMsgSystemPerfState, CMsgSystemPerfSettings
 export interface Perf {
